+++ application/views/jenis_pembayaran/index.php
@@ -34,72 +34,74 @@
       </div>
     <?php } ?>
     <!-- tambah data -->
-    <div class="row">
-      <div class="col-md-12">
-        <div class="card">
-          <div class="card-header">
-            <h5 class="card-title">Tambah Data Jenis Pembayaran</h5>
-            <div class="card-tools">
-              <button type="button" class="btn btn-tool" data-card-widget="collapse">
-                <i class="fas fa-minus"></i>
-              </button>
+    <?php if ($this->session->userdata('level') == 'admin') { ?>
+      <div class="row">
+        <div class="col-md-12">
+          <div class="card">
+            <div class="card-header">
+              <h5 class="card-title">Tambah Data Jenis Pembayaran</h5>
+              <div class="card-tools">
+                <button type="button" class="btn btn-tool" data-card-widget="collapse">
+                  <i class="fas fa-minus"></i>
+                </button>
+              </div>
             </div>
-          </div>
-          <!-- /.card-header -->
-          <div class="card-body">
-            <div class="row">
-              <div class="col-md-8">
-                <?= validation_errors(); ?>
-                <form action="<?= base_url() ?>DataJenisPembayaran/validation_form" method="post" accept-charset="utf-8">
-                  <div class="card-body">
-                    <!-- <div class="form-group">
+            <!-- /.card-header -->
+            <div class="card-body">
+              <div class="row">
+                <div class="col-md-8">
+                  <?= validation_errors(); ?>
+                  <form action="<?= base_url() ?>DataJenisPembayaran/validation_form" method="post" accept-charset="utf-8">
+                    <div class="card-body">
+                      <!-- <div class="form-group">
                       <label for="exampleInputEmail1">Kode Jenis Pembayaran</label>
                       <input type="text" class="form-control" id="exampleInputEmail1" name="kode_jenispembayaran">
                     </div> -->
-                    <div class="form-group">
-                      <label for="exampleInputPassword1">Nama Pembayaran Ujian</label>
-                      <select class="form-control" name="nama_pembayaran">
-                        <option>--Pilih Pembayaran--</option>
-                        <option value="UTS">UTS</option>
-                        <option value="UAS">UAS</option>
-                        <option value="UNBK">UNBK</option>
-                      </select>
-                    </div>
-                    <div class="form-group">
-                      <label for="exampleInputPassword1">Nominal Pembayaran</label>
-                      <input type="text" class="form-control" id="exampleInputPassword1" name="nominal">
-                    </div>
-                    <div class="form-group">
-                      <label for="exampleInputPassword1">Tahun Berlaku </label>
-                      <select class="form-control" name="kd_ta">
-                        <option>--Pilih Tahun--</option>
-                        <?php
+                      <div class="form-group">
+                        <label for="exampleInputPassword1">Nama Pembayaran Ujian</label>
+                        <select class="form-control" name="nama_pembayaran">
+                          <option>--Pilih Pembayaran--</option>
+                          <option value="UTS">UTS</option>
+                          <option value="UAS">UAS</option>
+                          <option value="UNBK">UNBK</option>
+                        </select>
+                      </div>
+                      <div class="form-group">
+                        <label for="exampleInputPassword1">Nominal Pembayaran</label>
+                        <input type="text" class="form-control" id="exampleInputPassword1" name="nominal">
+                      </div>
+                      <div class="form-group">
+                        <label for="exampleInputPassword1">Tahun Berlaku </label>
+                        <select class="form-control" name="kd_ta">
+                          <option>--Pilih Tahun--</option>
+                          <?php
 
-                        foreach ($tahunajaran as $row) { ?>
-                          <option value="<?= $row->kode_ta ?>" <?php echo set_select('kd_ta', $row->kode_ta); ?>><?= $row->tahun_ajaran ?></option>
-                        <?php } ?>
-                      </select>
+                          foreach ($tahunajaran as $row) { ?>
+                            <option value="<?= $row->kode_ta ?>" <?php echo set_select('kd_ta', $row->kode_ta); ?>><?= $row->tahun_ajaran ?></option>
+                          <?php } ?>
+                        </select>
 
+                      </div>
+                      <div class="form-group">
+                        <label for="exampleInputPassword1">Jumlah Pembayaran/Tahun </label>
+                        <input type="number" class="form-control" id="exampleInputPassword1" name="jumlah_pembayaran">
+                      </div>
+                      <input type="submit" name="save" class="btn btn-primary" value="Save">
                     </div>
-                    <div class="form-group">
-                      <label for="exampleInputPassword1">Jumlah Pembayaran/Tahun </label>
-                      <input type="number" class="form-control" id="exampleInputPassword1" name="jumlah_pembayaran">
-                    </div>
-                    <input type="submit" name="save" class="btn btn-primary" value="Save">
-                  </div>
-                  <!-- /.card-body -->
-                </form>
+                    <!-- /.card-body -->
+                  </form>
+                </div>
+                <!-- /.col -->
               </div>
-              <!-- /.col -->
+              <!-- /.row -->
             </div>
-            <!-- /.row -->
+            <!-- ./card-body -->
           </div>
-          <!-- ./card-body -->
+          <!-- /.card -->
         </div>
-        <!-- /.card -->
+        <!-- /.col -->
       </div>
-      <!-- /.col -->
-    </div>
+    <?php } ?>
     <!-- /.row -->
     <!-- list data -->
     <?= validation_errors(); ?>
@@ -118,7 +120,9 @@
                     <th>Nominal Jenis</th>
                     <th>Tahun</th>
                     <th>Jumlah Pembayaran Pertahun</th>
-                    <th>Action</th>
+                    <?php if ($this->session->userdata('level') == 'admin') { ?>
+                      <th>Action</th>
+                    <?php } ?>
                   </tr>
                 </thead>
                 <tbody>
@@ -132,12 +136,14 @@
                       <td><?= $row->nominal ?></td>
                       <td><?= $row->tahun_ajaran ?></td>
                       <td><?= $row->jumlah_pembayaran ?></td>
-                      <td>
-                        <div class="btn-group">
-                          <a href="<?= base_url() ?>DataJenisPembayaran/hapus/<?= $row->kode_jenispembayaran ?>" class="btn btn-danger" onclick="return confirm('Apakah Anada Akan Menghapus Data Ini ?')">Hapus</a>
-                          <a href="<?= base_url() ?>DataJenisPembayaran/ubah/<?= $row->kode_jenispembayaran ?>" class="btn btn-warning">update</a>
-                        </div>
-                      </td>
+                      <?php if ($this->session->userdata('level') == 'admin') { ?>
+                        <td>
+                          <div class="btn-group">
+                            <a href="<?= base_url() ?>DataJenisPembayaran/hapus/<?= $row->kode_jenispembayaran ?>" class="btn btn-danger" onclick="return confirm('Apakah Anada Akan Menghapus Data Ini ?')">Hapus</a>
+                            <a href="<?= base_url() ?>DataJenisPembayaran/ubah/<?= $row->kode_jenispembayaran ?>" class="btn btn-warning">update</a>
+                          </div>
+                        </td>
+                      <?php } ?>
                     </tr>
                   <?php
                     $no++;
