+++ application/views/jenis_spp/index.php
@@ -34,58 +34,60 @@
       </div>
     <?php } ?>
     <!-- tambah data -->
-    <div class="row">
-      <div class="col-md-12">
-        <div class="card">
-          <div class="card-header">
-            <h5 class="card-title">Tambah Data Jenis SPP</h5>
-            <div class="card-tools">
-              <button type="button" class="btn btn-tool" data-card-widget="collapse">
-                <i class="fas fa-minus"></i>
-              </button>
+    <?php if ($this->session->userdata('level') == 'admin') { ?>
+      <div class="row">
+        <div class="col-md-12">
+          <div class="card">
+            <div class="card-header">
+              <h5 class="card-title">Tambah Data Jenis SPP</h5>
+              <div class="card-tools">
+                <button type="button" class="btn btn-tool" data-card-widget="collapse">
+                  <i class="fas fa-minus"></i>
+                </button>
+              </div>
             </div>
-          </div>
-          <!-- /.card-header -->
-          <div class="card-body">
-            <div class="row">
-              <div class="col-md-8">
-                <?= validation_errors(); ?>
-                <form action="<?= base_url() ?>DataJenisSpp/validation_form" method="post" accept-charset="utf-8">
-                  <div class="card-body">
-                    <!-- <div class="form-group">
+            <!-- /.card-header -->
+            <div class="card-body">
+              <div class="row">
+                <div class="col-md-8">
+                  <?= validation_errors(); ?>
+                  <form action="<?= base_url() ?>DataJenisSpp/validation_form" method="post" accept-charset="utf-8">
+                    <div class="card-body">
+                      <!-- <div class="form-group">
                       <label for="exampleInputEmail1">Kode Jenis SPP</label>
                       <input type="text" class="form-control" id="exampleInputEmail1" name="kode_jenisspp">
                     </div> -->
 
 
-                    <div class="form-group">
-                      <label for="exampleInputPassword1">Tingkat</label>
-                      <input type="text" class="form-control" id="exampleInputPassword1" name="kategori">
-                    </div>
-                    <div class="form-group">
-                      <label for="exampleInputPassword1">Nominal Jenis</label>
-                      <input type="text" class="form-control" id="exampleInputPassword1" name="nominal_jenis">
-                    </div>
-                    <!-- <div class="form-group">
+                      <div class="form-group">
+                        <label for="exampleInputPassword1">Tingkat</label>
+                        <input type="text" class="form-control" id="exampleInputPassword1" name="kategori">
+                      </div>
+                      <div class="form-group">
+                        <label for="exampleInputPassword1">Nominal Jenis</label>
+                        <input type="text" class="form-control" id="exampleInputPassword1" name="nominal_jenis">
+                      </div>
+                      <!-- <div class="form-group">
                       <label for="exampleInputEmail1">Tahun</label>
                       <input type="text" class="form-control" id="exampleInputEmail1" name="tahun">
                     </div> -->
 
-                    <input type="submit" name="save" class="btn btn-primary" value="Save">
-                  </div>
-                  <!-- /.card-body -->
-                </form>
+                      <input type="submit" name="save" class="btn btn-primary" value="Save">
+                    </div>
+                    <!-- /.card-body -->
+                  </form>
+                </div>
+                <!-- /.col -->
               </div>
-              <!-- /.col -->
+              <!-- /.row -->
             </div>
-            <!-- /.row -->
+            <!-- ./card-body -->
           </div>
-          <!-- ./card-body -->
+          <!-- /.card -->
         </div>
-        <!-- /.card -->
+        <!-- /.col -->
       </div>
-      <!-- /.col -->
-    </div>
+    <?php } ?>
     <!-- /.row -->
     <!-- list data -->
     <div class="row">
@@ -100,8 +102,9 @@
                   <th>Kode Jenis SPP</th>
                   <th>Nominal Jenis</th>
                   <th>Tingkat</th>
-                  <!-- <th>Tahun</th> -->
-                  <th>Action</th>
+                  <?php if ($this->session->userdata('level') == 'admin') { ?>
+                    <th>Action</th>
+                  <?php } ?>
                 </tr>
               </thead>
               <tbody>
@@ -113,13 +116,14 @@
                     <td><?= $row->kode_jenisspp ?></td>
                     <td><?= $row->nominal_jenis ?></td>
                     <td><?= $row->kategori ?></td>
-                    <!-- <td><?= $row->tahun ?></td> -->
-                    <td>
-                      <div class="btn-group">
-                        <a href="<?= base_url() ?>DataJenisSpp/hapus/<?= $row->kode_jenisspp ?>" class="btn btn-danger" onclick="return confirm('Apakah Anada Akan Menghapus Data Ini ?')">Hapus</a>
-                        <a href="<?= base_url() ?>DataJenisSpp/ubah/<?= $row->kode_jenisspp ?>" class="btn btn-warning">update</a>
-                      </div>
-                    </td>
+                    <?php if ($this->session->userdata('level') == 'admin') { ?>
+                      <td>
+                        <div class="btn-group">
+                          <a href="<?= base_url() ?>DataJenisSpp/hapus/<?= $row->kode_jenisspp ?>" class="btn btn-danger" onclick="return confirm('Apakah Anada Akan Menghapus Data Ini ?')">Hapus</a>
+                          <a href="<?= base_url() ?>DataJenisSpp/ubah/<?= $row->kode_jenisspp ?>" class="btn btn-warning">update</a>
+                        </div>
+                      </td>
+                    <?php } ?>
                   </tr>
                 <?php
                   $no++;
