+++ application/views/jurusan/index.php
@@ -35,49 +35,55 @@
       </div>
     <?php } ?>
     <!-- tambah data -->
-    <div class="row">
-      <div class="col-md-12">
-        <div class="card">
-          <div class="card-header">
-            <h5 class="card-title">Tambah Data Jurusan</h5>
-            <div class="card-tools">
-              <button type="button" class="btn btn-tool" data-card-widget="collapse">
-                <i class="fas fa-minus"></i>
-              </button>
-            </div>
-          </div>
-          <!-- /.card-header -->
-          <div class="card-body">
-            <div class="row">
-              <div class="col-md-8">
-                <?= validation_errors(); ?>
-                <form action="<?= base_url() ?>DataJurusan/validation_form" method="post" accept-charset="utf-8">
-                  <div class="card-body">
-                    <div class="form-group">
-                      <label for="exampleInputEmail1">Kode Jurusan</label>
-                      <input type="text" class="form-control" id="exampleInputEmail1" name="kd_jur">
-                    </div>
-                    <div class="form-group">
-                      <label for="exampleInputPassword1">Nama Jurusan</label>
-                      <input type="text" class="form-control" id="exampleInputPassword1" name="nm_jur">
-                    </div>
-                    <input type="submit" name="save" class="btn btn-primary" value="Save">
-                  </div>
-                  <!-- /.card-body -->
-                </form>
+    <?php if ($this->session->userdata('level') == 'admin') { ?>
+      <div class="row">
+        <div class="col-md-12">
+          <div class="card">
+            <div class="card-header">
+
+              <h5 class="card-title">Tambah Data Jurusan</h5>
+              <div class="card-tools">
+                <button type="button" class="btn btn-tool" data-card-widget="collapse">
+                  <i class="fas fa-minus"></i>
+                </button>
               </div>
-              <!-- /.col -->
             </div>
-            <!-- /.row -->
+            <!-- /.card-header -->
+            <div class="card-body">
+              <div class="row">
+                <div class="col-md-8">
+                  <?= validation_errors(); ?>
+                  <form action="<?= base_url() ?>DataJurusan/validation_form" method="post" accept-charset="utf-8">
+                    <div class="card-body">
+                      <div class="form-group">
+                        <label for="exampleInputEmail1">Kode Jurusan</label>
+                        <input type="text" class="form-control" id="exampleInputEmail1" name="kd_jur">
+                      </div>
+                      <div class="form-group">
+                        <label for="exampleInputPassword1">Nama Jurusan</label>
+                        <input type="text" class="form-control" id="exampleInputPassword1" name="nm_jur">
+                      </div>
+                      <input type="submit" name="save" class="btn btn-primary" value="Save">
+                    </div>
+
+                    <!-- /.card-body -->
+                  </form>
+                </div>
+                <!-- /.col -->
+              </div>
+              <!-- /.row -->
+            </div>
+            <!-- ./card-body -->
           </div>
-          <!-- ./card-body -->
+          <!-- /.card -->
         </div>
-        <!-- /.card -->
+        <!-- /.col -->
       </div>
-      <!-- /.col -->
-    </div>
-    <!-- /.row -->
-    <!-- list data -->
+      <!-- /.row -->
+      <!-- list data -->
+    <?php
+    }
+    ?>
     <div class="row">
       <div class="col-12">
         <div class="card">
@@ -89,7 +95,11 @@
                   <th>No</th>
                   <th>Kode Jurusan</th>
                   <th>Jurusan</th>
-                  <th>Action</th>
+                  <?php if ($this->session->userdata('level') == 'admin') { ?>
+                    <th>Action</th>
+                  <?php
+                  }
+                  ?>
                 </tr>
               </thead>
               <tbody>
@@ -100,12 +110,16 @@
                     <td><?= $no ?></td>
                     <td><?= $row->kode_jurusan ?></td>
                     <td><?= $row->nama_jurusan ?></td>
-                    <td>
-                      <div class="btn-group">
-                        <a href="<?= base_url() ?>DataJurusan/hapus/<?= $row->kode_jurusan ?>" class="btn btn-danger" onclick="return confirm('yakin ?')">Hapus</a>
-                        <a href="<?= base_url() ?>DataJurusan/ubah/<?= $row->kode_jurusan ?>" class="btn btn-warning">update</a>
-                      </div>
-                    </td>
+                    <?php if ($this->session->userdata('level') == 'admin') { ?>
+                      <td>
+                        <div class="btn-group">
+                          <a href="<?= base_url() ?>DataJurusan/hapus/<?= $row->kode_jurusan ?>" class="btn btn-danger" onclick="return confirm('yakin ?')">Hapus</a>
+                          <a href="<?= base_url() ?>DataJurusan/ubah/<?= $row->kode_jurusan ?>" class="btn btn-warning">update</a>
+                        </div>
+                      </td>
+                    <?php
+                    }
+                    ?>
                   </tr>
                 <?php
                   $no++;
